checked for data older than 5 hrs and delete them

+++ server/scripts/cleanOldData.js
@@ -11,14 +11,17 @@ fs.readdir(directory, (err, files) => {
     fs.readFile(filePath, 'utf8', (err, data) => {
       if (err) throw err;
       const obj = JSON.parse(data);
+      if (obj.time >= Date.now() - (5 * 60 * 60 * 1000)) { //delete files older than 5 hrs
+        return;
+      }
       const imagePath = path.resolve(__dirname, obj.dir, obj.name);
       fs.unlink(imagePath, (err) => {
         if (err) throw err;
         console.log('image deleted', imagePath);
-        fs.unlink(filePath, (err) => {
-          if (err) throw err;
-          console.log('json deleted', filePath);
-        })
+        // fs.unlink(filePath, (err) => {
+        //   if (err) throw err;
+        //   console.log('json deleted', filePath);
+        // })
       })
     });
   })
